reorganized lambda functions for clarity
