+++ serverless.yml
@@ -59,12 +59,10 @@ package:
    - bin/**
    - lambda/handler.js
    - node_modules/cubic-spline/**
-  exclude:
-    - secrets.yml
 
 functions:
   calculator:
-    handler: lambda/handler.calculator
+    handler: lambda/calculatorLambda.calculator
     events:
       - http:
           path: version/calculator/{optionType}/{sensitivity}/{algorithm}
@@ -77,7 +75,7 @@ functions:
                 sensitivity: true
                 algorithm: true
   density:
-    handler: lambda/handler.density
+    handler: lambda/calculatorLambda.density
     events:
       - http:
           path: version/density/{densityType}
@@ -88,7 +86,7 @@ functions:
               paths:
                 densityType: true
   defaultParameters:
-    handler: lambda/handler.defaultParameters
+    handler: lambda/calculatorLambda.defaultParameters
     events:
       - http:
           path: version/parameters/parameter_ranges
@@ -96,7 +94,7 @@ functions:
           cors: true
 
   getExpirationDates:
-    handler: lambda/handler.getExpirationDates
+    handler: lambda/optionLambda.getExpirationDates
     events:
       - http:
           path: version/options/{ticker}/maturities
@@ -108,7 +106,7 @@ functions:
                 ticker: true
 
   getOptionPrices:
-    handler: lambda/handler.getOptionPrices
+    handler: lambda/optionLambda.getOptionPrices
     events:
       - http:
           path: version/options/{ticker}/prices/{asOfDate}
@@ -126,7 +124,7 @@ functions:
   calibrator:
     memorySize: 3008 # optional, in MB, default is 1024
     timeout: 30 # optional, in seconds, default is 6
-    handler: lambda/handler.calibrator
+    handler: lambda/calibratorLambda.calibrator
 
     events:
       - http:
